server/requirements.txt
mysqlclient
sqlalchemy
aiomysql
uvicorn
fastapi
email-validator
spotipy
python-dotenv
pandas
line-bot-sdk
